+++ StandBot/app/src/main/java/com/ispgaya/standbot/MostExpensive.java
@@ -19,7 +19,7 @@ public class MostExpensive {
 
     @SerializedName("preco")
     @Expose
-    private int preco;
+    private double preco;
 
     public String getMarca() {
         return marca;
@@ -45,11 +45,11 @@ public class MostExpensive {
         this.ano = ano;
     }
 
-    public int getPreco() {
+    public double getPreco() {
         return preco;
     }
 
-    public void setPreco(int preco) {
+    public void setPreco(double preco) {
         this.preco = preco;
     }
 }
